Add forward path intercept and changed the database config for proxy +refactored compileTemplate
Refactored createTemplate to make more generic and offer better options.  Added ability to add slugger and changed the proxy config so all proxy settings are under one property.  Also, now has ability to fire a sequence of proxies(just in case, ... well, guess this is not 100% useful but why not?)